refactor(pptx): déplacer SlideChunker vers la couche application
SlideChunker n'a aucune dépendance externe (regex uniquement) et ne doit pas
résider dans infrastructure. Déplacé vers application/services/ et test_slide_chunker
vers tests/unit/application/services/ pour respecter la symétrie source/tests.
